Split the diff view types out of types.ts to get main back under the lint cap
`main` was red: `erun-ui/frontend/src/types.ts` had grown to 509 counted lines
against eslint's `max-lines: 500`. That fails `erun-ui/build.sh`, so the desktop
did not build at all, and it fails `playwright/run.sh`, which runs
typecheck + lint + format:check before any spec — so every environment's
Playwright run was blocked too, not just the build.

No single change caused it. Several merges each added a field or two to the same
file until it crossed the line, which is why it passed every PR's own gate and
only broke once they were all in.

Moving the eight diff view types (`DiffResult`, `DiffSummary`, `DiffFile`,
`DiffHunk`, `DiffLine`, `DiffTreeNode`, `DiffReviewBase`, `DiffCommit`) into
`diffTypes.ts`. They are self-contained — the split needs no imports back into
`types.ts` — and `types.ts` re-exports the module, so every existing
`from './types'` import keeps resolving with no call-site churn.

Extracting rather than raising the cap, matching how the same class of limit was
handled for `cyclop` in #1269.

    types.ts 692 -> 625 lines, diffTypes.ts 73

Verified after regenerating the gitignored wailsjs bindings (without them every
`wailsjs` import fails and cascades ~365 phantom lint errors):
eslint clean, tsc clean, prettier --check clean, frontend units 129/129,
erun-ui go tests green, and erun-ui/build.sh exits 0 with a desktop binary.
