+++ erun-ui/frontend/src/types.ts
@@ -620,73 +620,6 @@ export interface AgentOutputsList {
   truncated: boolean;
 }
 
-export interface DiffResult {
-  workingDirectory?: string;
-  rawDiff: string;
-  summary: DiffSummary;
-  files?: DiffFile[];
-  tree?: DiffTreeNode[];
-  reviewBase?: DiffReviewBase;
-  reviewCommits?: DiffCommit[];
-  scope?: 'current' | 'commit' | 'all';
-  selectedCommit?: string;
-  includesWorktree?: boolean;
-}
-
-export interface DiffSummary {
-  fileCount: number;
-  additions: number;
-  deletions: number;
-}
-
-export interface DiffFile {
-  path: string;
-  oldPath?: string;
-  newPath?: string;
-  status: string;
-  additions: number;
-  deletions: number;
-  binary?: boolean;
-  hunks?: DiffHunk[];
-}
-
-export interface DiffHunk {
-  header: string;
-  oldStart: number;
-  oldLines: number;
-  newStart: number;
-  newLines: number;
-  lines?: DiffLine[];
-}
-
-export interface DiffLine {
-  kind: 'context' | 'add' | 'delete' | 'meta';
-  content: string;
-  oldLine?: number;
-  newLine?: number;
-}
-
-export interface DiffTreeNode {
-  name: string;
-  path: string;
-  parentPath?: string;
-  type: 'directory' | 'file';
-  depth: number;
-  status?: string;
-  additions?: number;
-  deletions?: number;
-}
-
-export interface DiffReviewBase {
-  branch?: string;
-  commit?: string;
-  shortCommit?: string;
-}
-
-export interface DiffCommit {
-  hash: string;
-  shortHash: string;
-  subject: string;
-  author: string;
-  date: string;
-}
+// Diff view types live in ./diffTypes to keep this file under eslint's max-lines
+// cap; re-exported here so `from './types'` keeps resolving them.
+export * from './diffTypes';
